backend/eloquence-backend/requirements.txt
livekit-agents>=1.0.0
livekit-api>=1.0.0
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
websockets>=11.0.3
requests>=2.28.0
numpy>=1.22.0
asyncpg>=0.20.0
httpx>=0.20.0
python-multipart>=0.0.5
sqlalchemy>=1.4.0 # Ou la version appropriée
sqlalchemy-utils>=0.38.0 # Si utilisé
asyncpg>=0.27.0 # Si utilisé avec asyncpg
aiosqlite # Pour SQLite asynchrone en mode test
torchaudio # Pour le service VAD
celery # Pour les tâches asynchrones
PyYAML>=6.0 # Ou la version appropriée
pydantic-settings>=2.0.0 # Ou la version appropriée
soundfile # Ou la version appropriée
faster-whisper # Ou la version appropriée
redis # Ou la version appropriée
sqlalchemy>=1.4.0 # Ou la version appropriée
sqlalchemy-utils>=0.38.0 # Si utilisé
asyncpg>=0.27.0 # Si utilisé avec asyncpg
